requirements.txt
azure_storage==0.37.0
beautifulsoup4==4.11.1
discord.py==2.0.1
lxml==4.9.1
numpy==1.21.2
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.26.0
selenium==4.5.0
